update calculating overall price for cart

+++ app/src/main/java/com/example/appbanlinhkien30/util/Convert.java
@@ -5,12 +5,14 @@ import java.util.Currency;
 import java.util.Locale;
 
 public class Convert {
+    //Chuyển đổi chuỗi tiền tệ sang số thực
     public static String convertCurrencyStringToNumber(String currencyString) {
         String numericString = currencyString.replaceAll("[^0-9]", "");
-        double numericValue = Integer.parseInt(numericString);
+        double numericValue = Double.parseDouble(numericString);
         return String.valueOf(numericValue);
     }
 
+    //Chuyển đổi số thực sang chuỗi tiền tệ
     public static String convertNumberToCurrencyString(double numericValue){
             Locale locale = new Locale("vi", "VN");
             Currency currency = Currency.getInstance("VND");
